+++ packages/react-native/examples/ReactNativeWallet/App.tsx
@@ -1,0 +1,19 @@
+import { ReactNativeCrypto } from "@lightsparkdev/react-native";
+import {
+  JwtAuthProvider,
+  LightsparkClientProvider,
+} from "@lightsparkdev/react-wallet";
+import PageContainer from "./PageContainer";
+
+export default function App() {
+  return (
+    <LightsparkClientProvider
+      serverUrl="api.dev.dev.sparkinfra.net"
+      customCryptoImpl={ReactNativeCrypto}
+    >
+      <JwtAuthProvider useLocalStorage>
+        <PageContainer />
+      </JwtAuthProvider>
+    </LightsparkClientProvider>
+  );
+}