+++ packages/react-wallet/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @lightsparkdev/react-wallet
 
+## 0.3.2
+
+### Patch Changes
+
+- Refactor internals to allow for a custom react native crypto implementation to be injected into the LightsparkClient.
+- Updated dependencies
+  - @lightsparkdev/core@0.2.2
+  - @lightsparkdev/wallet-sdk@0.3.2
+
 ## 0.3.1
 
 ### Patch Changes